Refactor project structure and update Dockerfile
Refactored the project structure to enhance the modularity and improve the organization. The project relies more on SSI (Server Side Inclusions), hence, scripts previously added to the index.html file directly are moved to a separate _js.html file. All PHP includes are changed to HTML includes due to the increased use of SSI over PHP.

The Dockerfile is updated to use layered building and simplify the setup process. It now uses pre-built Go-based supervisord instead of PHP FPM, reducing the complexity of the Dockerfile and improving build performance. A memcached key is now passed as a header to PHP.

README.md file is updated for simplicity; changed from using an absolute URL to a relative path for the workflow diagram, and updated the docker build command.