Remove redundant line in Dockerfile
The final Dockerfile had an extraneous line left in it from previous iterations of the file that was no longer needed. This line has been removed to simplify reading and understanding the file for future maintainers and/or Docker users. It has no functional impact, but is a matter of cleanliness and proper commenting.